# tests/contract/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "GraphNodeDetailView": {
        "properties": {
          "node": {
            "$ref": "#/components/schemas/GraphNodeView"
          },
          "overlay": {
            "$ref": "#/components/schemas/GraphOverlayView"
          },
          "prerequisite_node_ids": {
            "items": {
              "type": "string"
            },
            "title": "Prerequisite Node Ids",
            "type": "array"
          },
          "successor_node_ids": {
            "items": {
              "type": "string"
            },
            "title": "Successor Node Ids",
            "type": "array"
          }
        },
        "required": [
          "node",
          "overlay",
          "prerequisite_node_ids",
          "successor_node_ids"
        ],
        "title": "GraphNodeDetailView",
        "type": "object"
      },
      "GraphNodeView": {
        "properties": {
          "group_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Group Key"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "node_id": {
            "pattern": "^n\\d{3,}$",
            "title": "Node Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "title",
          "group_key",
          "metadata"
        ],
        "title": "GraphNodeView",
        "type": "object"
      },
      "GraphOverlayView": {
        "properties": {
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "learning",
                  "proficient",
                  "expert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "status_source": {
            "anyOf": [
              {
                "enum": [
                  "user",
                  "summary_memory",
                  "system_recompute"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Source"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "node_id",
          "status",
          "version",
          "status_source",
          "updated_at"
        ],
        "title": "GraphOverlayView",
        "type": "object"
      },
      "GraphStateChangeView": {
        "properties": {
          "after_status": {
            "anyOf": [
              {
                "enum": [
                  "learning",
                  "proficient",
                  "expert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Status"
          },
          "after_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Version"
          },
          "before_status": {
            "anyOf": [
              {
                "enum": [
                  "learning",
                  "proficient",
                  "expert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Status"
          },
          "before_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Version"
          },
          "changed_at": {
            "format": "date-time",
            "title": "Changed At",
            "type": "string"
          },
          "node_id": {
            "pattern": "^n\\d{3,}$",
            "title": "Node Id",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Reason Codes",
            "type": "array"
          },
          "source_type": {
            "enum": [
              "user",
              "summary_memory",
              "system_recompute"
            ],
            "title": "Source Type",
            "type": "string"
          }
        },
        "required": [
          "node_id",
          "before_status",
          "after_status",
          "before_version",
          "after_version",
          "source_type",
          "changed_at"
        ],
        "title": "GraphStateChangeView",
        "type": "object"
      },
      "GraphStateExplanation": {
        "properties": {
          "changed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Changed At"
          },
          "current_status": {
            "anyOf": [
              {
                "enum": [
                  "learning",
                  "proficient",
                  "expert"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Status"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Evidence Refs",
            "type": "array"
          },
          "explanation_available": {
            "title": "Explanation Available",
            "type": "boolean"
          },
          "node_id": {
            "title": "Node Id",
            "type": "string"
          },
          "reason_codes": {
            "items": {
              "type": "string"
            },
            "title": "Reason Codes",
            "type": "array"
          },
          "source_memory_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Memory Id"
          },
          "source_memory_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Memory Version"
          },
          "source_type": {
            "anyOf": [
              {
                "enum": [
                  "user",
                  "summary_memory",
                  "system_recompute"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Type"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          }
        },
        "required": [
          "node_id",
          "current_status",
          "explanation_available",
          "summary",
          "reason_codes",
          "source_type",
          "source_memory_id",
          "source_memory_version",
          "evidence_refs",
          "changed_at"
        ],
        "title": "GraphStateExplanation",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryOperationResult": {
        "properties": {
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PublicError"
              },
              {
                "type": "null"
              }
            ]
          },
          "graph_state_changes": {
            "items": {
              "$ref": "#/components/schemas/GraphStateChangeView"
            },
            "title": "Graph State Changes",
            "type": "array"
          },
          "mutations": {
            "items": {
              "$ref": "#/components/schemas/MutationResult"
            },
            "title": "Mutations",
            "type": "array"
          },
          "operation_id": {
            "format": "uuid",
            "title": "Operation Id",
            "type": "string"
          },
          "operation_type": {
            "enum": [
              "conversation_evidence",
              "activity_evidence",
              "correct_memory",
              "forget_memory",
              "restore_memory",
              "override_learner_profile",
              "review_candidate",
              "set_graph_state",
              "project_summary_to_graph",
              "rebuild_index",
              "verify_checksums",
              "purge_tombstones",
              "cleanup_orphan_versions",
              "cleanup_checkpoints",
              "purge_account_memory"
            ],
            "title": "Operation Type",
            "type": "string"
          },
          "review_candidate_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Review Candidate Ids",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "retry_wait",
              "succeeded",
              "needs_review",
              "dead_letter",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "operation_id",
          "status",
          "operation_type",
          "created_at",
          "updated_at"
        ],
        "title": "MemoryOperationResult",
        "type": "object"
      },
      "MutationResult": {
        "properties": {
          "action": {
            "enum": [
              "create",
              "merge",
              "replace",
              "append_evidence",
              "forget",
              "restore"
            ],
            "title": "Action",
            "type": "string"
          },
          "after_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "After Version"
          },
          "before_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Before Version"
          },
          "memory_id": {
            "title": "Memory Id",
            "type": "string"
          },
          "mutation_id": {
            "format": "uuid",
            "title": "Mutation Id",
            "type": "string"
          }
        },
        "required": [
          "mutation_id",
          "memory_id",
          "action",
          "before_version",
          "after_version"
        ],
        "title": "MutationResult",
        "type": "object"
      },
...
      "PublicError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Field"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "trace_id": {
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "retryable",
          "trace_id"
        ],
        "title": "PublicError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "_GraphStatePutBody": {
        "additionalProperties": false,
        "description": "PUT state 原始 body：action 先按字符串接收，以便把 expert 映射到专用错误码。\n\n字段集合与 GraphStatePutRequest 一致（extra=\"forbid\"，§6.4）。",
        "properties": {
          "action": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Action",
            "type": "string"
          },
          "expected_version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          }
        },
        "required": [
          "action"
        ],
        "title": "_GraphStatePutBody",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "MemoryManagerGraph API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/knowledge-graph/me/nodes/{node_id}": {
      "get": {
        "operationId": "get_my_node_detail_api_v1_knowledge_graph_me_nodes__node_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "pattern": "^n\\d{3,}$",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphNodeDetailView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get My Node Detail",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
    "/api/v1/knowledge-graph/me/nodes/{node_id}/explanation": {
      "get": {
        "description": "由最近一次 graph_state_audit 生成受控解释（§19.5）。",
        "operationId": "get_node_explanation_api_v1_knowledge_graph_me_nodes__node_id__explanation_get",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "pattern": "^n\\d{3,}$",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GraphStateExplanation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Node Explanation",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
    "/api/v1/knowledge-graph/me/nodes/{node_id}/state": {
      "delete": {
        "description": "clear：有 Overlay 缺版本 422；无 Overlay 可省略版本并幂等 no_change（§19.5）。",
        "operationId": "delete_node_state_api_v1_knowledge_graph_me_nodes__node_id__state_delete",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "pattern": "^n\\d{3,}$",
              "title": "Node Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Expected Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Node State",
        "tags": [
          "knowledge-graph"
        ]
      },
      "put": {
        "operationId": "put_node_state_api_v1_knowledge_graph_me_nodes__node_id__state_put",
        "parameters": [
          {
            "in": "path",
            "name": "node_id",
            "required": true,
            "schema": {
              "pattern": "^n\\d{3,}$",
              "title": "Node Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/_GraphStatePutBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryOperationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Node State",
        "tags": [
          "knowledge-graph"
        ]
      }
    },
...
  }
}